Deleted ViewSimpleTest and pasted implementation to LoginView
